+++ wemake_python_styleguide/visitors/presets/tokens.py
@@ -1,18 +1,16 @@
 # -*- coding: utf-8 -*-
 
-from wemake_python_styleguide.visitors.tokenize.comments import (
-    WrongCommentVisitor,
-)
-from wemake_python_styleguide.visitors.tokenize.keywords import (
-    WrongKeywordTokenVisitor,
-)
-from wemake_python_styleguide.visitors.tokenize.primitives import (
-    WrongPrimitivesVisitor,
+from wemake_python_styleguide.visitors.tokenize import (
+    comments,
+    keywords,
+    primitives,
+    statements,
 )
 
 #: Used to store all token related visitors to be later passed to checker:
 TOKENS_PRESET = (
-    WrongCommentVisitor,
-    WrongKeywordTokenVisitor,
-    WrongPrimitivesVisitor,
+    comments.WrongCommentVisitor,
+    keywords.WrongKeywordTokenVisitor,
+    primitives.WrongPrimitivesVisitor,
+    statements.ExtraIndentationVisitor,
 )